leecode 简单题目提交

+++ design/suanfa/src/main/java/数组集合AA/数据流中的中位数/Solution.java
@@ -20,7 +20,7 @@ public class Solution {
     }
 
     public Double GetMedian() {
-        return max.size() == min.size() ? (max.peek() + min.peek()) / 2.0 : max.peek();
+        return (Double) (max.size() == min.size() ? (max.peek() + min.peek()) / 2.0 : max.peek());
     }
